Changes to enable simple building of Java support on Linux. Includes minor code changes to remove a few gcc compiler warnings.
git-svn-id: https://openbiometrics.svn.sourceforge.net/svnroot/openbiometrics/OpenEBTS@180 53fe7bc9-f1be-4a37-98cb-3b35ff32843c

+++ OpenEBTS/IWVerification.cpp
@@ -1047,10 +1047,10 @@ void CIWVerification::DebugOutputVerification()
 
 		// Output all TOTs for this transaction list
 		sMsg.Empty();
-		for (j = 0; j < pTrans->m_TOTArray.size(); j++)
+		for (j = 0; j < (unsigned int)pTrans->m_TOTArray.size(); j++)
 		{
 			sMsg += pTrans->m_TOTArray[j];
-			if (j != (long)pTrans->m_TOTArray.size()-1)
+			if (j != (unsigned int)pTrans->m_TOTArray.size()-1)
 			{
 				sMsg += _T(", ");
 			}
